CStdFile: Use size_t instead of int for some size parameters
This avoids casting size_t * to int * in C4Log::GetLogSection

+++ src/C4MusicSystem.cpp
@@ -370,7 +370,7 @@ void C4MusicSystem::LoadMoreMusic()
 {
 	// Read MoreMusic.txt file or cancel if not present
 	CStdFile MoreMusicFile;
-	std::uint8_t *fileContentsTmp; int size;
+	std::uint8_t *fileContentsTmp; size_t size;
 	if (!MoreMusicFile.Load(Config.AtExePath(C4CFN_MoreMusic), &fileContentsTmp, &size)) return;
 	std::unique_ptr<std::uint8_t[]> fileContents(fileContentsTmp);
